+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Technical - Add more tests consistency.
 - Technical - Reorganize analysis services.
 - Technical - Add more tests on the dumper service.
+- Technical - Add tests on the table constraints getter service.
 
 ### Fixed
 - MySQL - Handle BIT(1) boolean columns and handle the buffer<01> value returned on join queries.